@@ -25,16 +25,11 @@ function requestBody(request) {
     throw new Error(`Unexpected request.body type: ${typeof request.body}`);
 }
 
-module.exports = (request, options) => {
+module.exports = (request) => {
     const method = request.method;
     const query = request.query;
     const headers = requestHeaders(request);
     const body = requestBody(request);
-
-    if (typeof options.requestId === 'string' && options.requestId.length > 0) {
-        const header = options.requestId.toLowerCase();
-        headers[header] = headers[header] || event.requestContext.requestId;
-    }
 
     const req = new Request({
         method,
