@@ -1,6 +1,3 @@
-/// <reference types="aws-lambda" />
-import { Context, HttpRequest } from '@azure/functions';
-
 declare namespace ServerlessHttp {
   export interface FrameworkApplication {
     callback: Function;
@@ -17,24 +14,27 @@ declare namespace ServerlessHttp {
    * Handler-compatible function or application.
    */
   export type Application = Function | Partial<FrameworkApplication>;
+  export type Result = Function | Partial<FrameworkApplication>;
 
+  export type Options = {
+    provider?: 'aws' | 'azure'
+    requestId?: string,
+    request?: Object | Function,
+    response?: Object | Function,
+    binary?: boolean | Function | string | string[]
+  }
   /**
    * AWS Lambda APIGatewayProxyHandler-like handler.
    */
   export type Handler = (
-    event: AWSLambda.APIGatewayProxyEvent | AWSLambda.APIGatewayProxyEventV2,
-    context: AWSLambda.Context
-  ) => Promise<AWSLambda.APIGatewayProxyResult | AWSLambda.APIGatewayProxyStructuredResultV2>;
-
-  export type AzureHandler = (
-    context: Context,
-    req: HttpRequest
-  ) => Promise<{ [key: string]: any }>;
+    event: Object,
+    context: Object
+  ) => Promise<Object>;
 }
 
 /**
  * Wraps the application into a Lambda APIGatewayProxyHandler-like handler.
  */
-declare function ServerlessHttp(application: ServerlessHttp.Application, options?: any): ServerlessHttp.Handler | ServerlessHttp.AzureHandler;
+declare function ServerlessHttp(application: ServerlessHttp.Application, options?: ServerlessHttp.Options): ServerlessHttp.Handler;
 
 export = ServerlessHttp;
